chore: Promote some integration tests to js_unit_tests
