--- cli/tests/testdata/workers/test.ts.out
@@ -1,3 +1,0 @@
-[WILDCARD]
-ok | [WILDCARD] passed | 0 failed ([WILDCARD])
-
