--- cli/tests/unit/worker_types.ts
@@ -1,14 +1,0 @@
-// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-import { assert } from "./test_util.ts";
-
-Deno.test(
-  { permissions: { read: true } },
-  function utimeSyncFileSuccess() {
-    const w = new Worker(
-      import.meta.resolve("../testdata/workers/worker_types.ts"),
-      { type: "module" },
-    );
-    assert(w);
-    w.terminate();
-  },
-);
